third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-time/timeline-timer.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/timeline-test.js"></script>
<script>

function performActions()
{
    var callback;
    var promise = new Promise((fulfill) => callback = fulfill);
    var timerOne = setTimeout("1 + 1", 10);
    var timerTwo = setInterval(intervalTimerWork, 20);
    var iteration = 0;

    function intervalTimerWork()
    {
        if (++iteration < 2)
            return;
        clearInterval(timerTwo);
        callback();
    }
    return promise;
}

function test() {
  UI.panels.timeline._disableCaptureJSProfileSetting.set(true);
  PerformanceTestRunner.invokeAsyncWithTimeline('performActions', finish);

  function finish() {
    PerformanceTestRunner.printTimelineRecordsWithDetails('TimerInstall');
    PerformanceTestRunner.printTimelineRecordsWithDetails('TimerFire');
    PerformanceTestRunner.printTimelineRecordsWithDetails('TimerRemove');
    PerformanceTestRunner.printTimelineRecords('FunctionCall');
    PerformanceTestRunner.printTimelineRecordsWithDetails('EvaluateScript');
    TestRunner.completeTest();
  }
}

if (!window.testRunner)
    setTimeout(performActions, 2000);

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline events for Timers
</p>
</body>
</html>
